Add IgnoreServiceErrorAttribute to Azure.Core.TestFramework
Can be used for intermittent service errors. Ported loosely from sdk/keyvault.
