+++ sdk/core/Azure.Core.TestFramework/src/RecordedTestBaseOfTEnvironment.cs
@@ -7,6 +7,7 @@ using NUnit.Framework;
 
 namespace Azure.Core.TestFramework
 {
+    [NonParallelizable]
 #pragma warning disable SA1649 // File name should match first type name
     public abstract class RecordedTestBase<TEnvironment> : RecordedTestBase where TEnvironment : TestEnvironment, new()
 #pragma warning restore SA1649 // File name should match first type name
