+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/KeyClientLiveTests.SecureKeyRelease.cs
@@ -14,8 +14,8 @@ namespace Azure.Security.KeyVault.Keys.Tests
     public partial class KeyClientLiveTests
     {
         [Test]
-        [AttestationMayFail] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
-        [PartiallyDeployed] // TODO: Remove once SKR is deployed to sovereign clouds.
+        [IgnoreServiceError(403, "Forbidden", Message = "Target environment attestation statement cannot be verified.")] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
+        [IgnoreServiceError(400, "BadParameter")] // TODO: Remove once SKR is deployed to sovereign clouds.
         [PremiumOnly]
         [ServiceVersion(Min = KeyClientOptions.ServiceVersion.V7_3)]
         public async Task ReleaseCreatedKey()
@@ -42,8 +42,8 @@ namespace Azure.Security.KeyVault.Keys.Tests
         }
 
         [Test]
-        [AttestationMayFail] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
-        [PartiallyDeployed] // TODO: Remove once SKR is deployed to sovereign clouds.
+        [IgnoreServiceError(403, "Forbidden", Message = "Target environment attestation statement cannot be verified.")] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
+        [IgnoreServiceError(400, "BadParameter")] // TODO: Remove once SKR is deployed to sovereign clouds.
         [PremiumOnly]
         [ServiceVersion(Min = KeyClientOptions.ServiceVersion.V7_3)]
         public async Task ReleaseUpdatedKey()
@@ -80,8 +80,8 @@ namespace Azure.Security.KeyVault.Keys.Tests
         }
 
         [Test]
-        [AttestationMayFail] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
-        [PartiallyDeployed] // TODO: Remove once SKR is deployed to sovereign clouds.
+        [IgnoreServiceError(403, "Forbidden", Message = "Target environment attestation statement cannot be verified.")] // TODO: Remove once the attestation issue is resolved: https://github.com/Azure/azure-sdk-for-net/issues/27957
+        [IgnoreServiceError(400, "BadParameter")] // TODO: Remove once SKR is deployed to sovereign clouds.
         [PremiumOnly]
         [ServiceVersion(Min = KeyClientOptions.ServiceVersion.V7_3)]
         public async Task UpdateReleasePolicy([Values] bool immutable)
