+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/KeysTestBase.cs
@@ -17,7 +17,6 @@ namespace Azure.Security.KeyVault.Keys.Tests
         KeyClientOptions.ServiceVersion.V7_1,
         KeyClientOptions.ServiceVersion.V7_2,
         KeyClientOptions.ServiceVersion.V7_3)]
-    [NonParallelizable]
     public abstract class KeysTestBase : RecordedTestBase<KeyVaultTestEnvironment>
     {
         protected TimeSpan PollingInterval => Recording.Mode == RecordedTestMode.Playback
