+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/SampleFixture.cs
@@ -7,8 +7,6 @@ using NUnit.Framework;
 
 namespace Azure.Security.KeyVault.Keys.Samples
 {
-    [LiveOnly]
-    [NonParallelizable]
     public class SampleFixture: SamplesBase<KeyVaultTestEnvironment>
     {
         private KeyVaultTestEventListener _listener;
