+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/SessionRecords/KeyClientLiveTests/ExportableNotSupportedThrowsInconclusiveAsync.json
@@ -1,0 +1,7 @@
+﻿{
+    "Entries": [],
+    "Variables": {
+        "AZURE_KEYVAULT_URL": "https://heathskeyvault.vault.azure.net/",
+        "RandomSeed": "1807643676"
+    }
+}
