+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/tests/SampleFixture.cs
@@ -7,8 +7,6 @@ using NUnit.Framework;
 
 namespace Azure.Security.KeyVault.Secrets.Samples
 {
-    [LiveOnly]
-    [NonParallelizable]
     public class SampleFixture: SamplesBase<KeyVaultTestEnvironment>
     {
         private KeyVaultTestEventListener _listener;
