+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/tests/SecretsTestBase.cs
@@ -18,7 +18,6 @@ namespace Azure.Security.KeyVault.Secrets.Tests
         SecretClientOptions.ServiceVersion.V7_1,
         SecretClientOptions.ServiceVersion.V7_2,
         SecretClientOptions.ServiceVersion.V7_3)]
-    [NonParallelizable]
     public abstract class SecretsTestBase : RecordedTestBase<KeyVaultTestEnvironment>
     {
         protected TimeSpan PollingInterval => Recording.Mode == RecordedTestMode.Playback
